remove open weather data source impl
